requirements.txt
# FastAPI и веб-сервер
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP клиент и переменные окружения
aiohttp
python-dotenv

# OpenAI/OpenRouter интеграция
openai

# Логирование
logging
